feat: allow setting python version in build script
Jupyterbooks does not yet support sphinx v4, and it looks like sphinx v3
has some issues running with python3.10.

For example: https://github.com/sphinx-doc/sphinx/issues/9512

was fixed in sphinx 4, but not in 3 from the looks of it.

So, for the time being, we stick to python <= 3.9.

So, for example, to create a new virtualenv using 3.9, use:

```
PYTHON=python3.9 ./build-helper.sh -c
```