feat(404): update urls to ensure scripts etc are loaded
When generated by jupyter-book, the URLs to the scripts are generated
relative to the location of the markdown source file. Since the 404.md
file is in the source root, the URLs to the `_static` folder is
generated as `"_static"`, since it's also in the source root.

However, when the 404.html page is used in GitHub pages, we're not
redirected to docs.neuroml.org/404.html, instead the contents of the
page are loaded wherever the 404 occurs, e.g.,
docs.neuroml.org/something/something.html. In this case, the scripts do
not load correctly because the path to the `_static` folder is
incorrect.

A simple workaround is to provide an absolute path to the `_static`
folder rather than a relative one, which is what we do here now.

A CAVEAT here is that when building the book locally, we'll still be
pointing to the scripts online. This should usually be fine, but is good
to keep in mind.